fixed norminette, all ok
