+++ push_swap.h
@@ -6,7 +6,7 @@
 /*   By: rmiranda <rmiranda@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/22 17:20:42 by rmiranda          #+#    #+#             */
-/*   Updated: 2023/02/23 05:14:20 by rmiranda         ###   ########.fr       */
+/*   Updated: 2023/02/23 05:57:25 by rmiranda         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -56,6 +56,7 @@ void	sort_b_size_3(t_sort_info sort_info);
 void	sort_size_5(t_sort_info sort_info);
 void	merge_sort(t_sort_info sort_info);
 int		push_and_count(t_sort_info sort_info);
+int		find_middle_value(t_sort_info sort_info, int *error);
 
 // PUSH SWAP COMMANDS
 int		pusw_pa(t_node **stack_a, t_node **stack_b, int output_command);
