+++ sort_functions/merge_sort.c
@@ -6,7 +6,7 @@
 /*   By: rmiranda <rmiranda@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/02/17 03:29:15 by rmiranda          #+#    #+#             */
-/*   Updated: 2023/02/23 05:12:54 by rmiranda         ###   ########.fr       */
+/*   Updated: 2023/02/23 06:19:43 by rmiranda         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
