+++ sort_functions/simple_sort.c
@@ -6,7 +6,7 @@
 /*   By: rmiranda <rmiranda@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/02/22 21:51:34 by rmiranda          #+#    #+#             */
-/*   Updated: 2023/02/23 05:11:12 by rmiranda         ###   ########.fr       */
+/*   Updated: 2023/02/23 05:37:40 by rmiranda         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -42,7 +42,7 @@ void	sort_a_size_3(t_sort_info sort_info)
 	stack_a = sort_info.stack_a;
 	if (stack_a[0]->value > stack_a[0]->next->value
 		&& stack_a[0]->value > stack_a[0]->next->next->value)
-			pusw_ra(stack_a, OUTPUT_COMMAND);
+		pusw_ra(stack_a, OUTPUT_COMMAND);
 	else if (stack_a[0]->next->value > stack_a[0]->next->next->value)
 	{
 		pusw_ra(stack_a, OUTPUT_COMMAND);
@@ -59,7 +59,7 @@ void	sort_b_size_3(t_sort_info sort_info)
 	stack_b = sort_info.stack_b;
 	if (stack_b[0]->value < stack_b[0]->next->value
 		&& stack_b[0]->value < stack_b[0]->next->next->value)
-			pusw_rb(stack_b, OUTPUT_COMMAND);
+		pusw_rb(stack_b, OUTPUT_COMMAND);
 	else if (stack_b[0]->next->value < stack_b[0]->next->next->value)
 	{
 		pusw_rb(stack_b, OUTPUT_COMMAND);
